# openapi.yaml
openapi: "3.1.0"
info:
  title: SCS-RG — Smart Campus Safety & Response Grid API
  version: "1.0.0"
  description: |
    Backend API for Multi-Hazard Smart Campus Safety & Response Grid.
    Handles sensor ingestion, risk fusion, incident lifecycle, priority ranking,
    and actuator commands for campus zone monitoring (RoboFusion Track B).
servers:
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: scs_session
    ZoneApiKey:
      type: apiKey
      in: header
      name: x-zone-api-key
  schemas:
    SafetyState:
      type: string
      enum: [SAFE, WARNING, CRITICAL, OFFLINE, NOT_CONFIGURED]
...
    HazardType:
      type: string
      enum: [FIRE, GAS, FLOOD, OCCUPANCY, NONE]
...
    ActuatorCommand:
      type: object
      properties:
        command_id: { type: string }
        command_version: { type: integer }
        safety_state: { $ref: "#/components/schemas/SafetyState" }
        led: { type: string, enum: [GREEN, YELLOW, RED, BLUE, OFF] }
        buzzer: { type: boolean }
        relay_cutoff: { type: boolean }
        issued_at: { type: string, format: date-time }
security:
  - SessionCookie: []
paths:
...
  /api/v1/incidents:
    get:
      summary: List incidents with optional filters
      tags: [Incidents]
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [active, all, resolved, OPEN, ACKNOWLEDGED, RESOLVED] }
        - in: query
          name: zoneId
          schema: { type: string }
        - in: query
          name: zoneCode
          schema: { type: string }
        - in: query
          name: hazard
          schema: { $ref: "#/components/schemas/HazardType" }
        - in: query
          name: from
          schema: { type: string, format: date-time }
          description: Filter incidents opened after this timestamp
        - in: query
          name: to
          schema: { type: string, format: date-time }
          description: Filter incidents opened before this timestamp
      responses:
        "200": { description: Incident list }
        "401": { description: Not authenticated }
...
  /api/v1/commands/{zoneCode}:
    get:
      summary: Get latest actuator command for a zone node
      security:
        - ZoneApiKey: []
      tags: [Commands]
      parameters:
        - in: path
          name: zoneCode
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Current command
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActuatorCommand" }
        "401": { description: Invalid zone API key }
    post:
      summary: Zone node acknowledges command applied
      security:
        - ZoneApiKey: []
      tags: [Commands]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [command_id]
              properties:
                command_id: { type: string }
      responses:
        "200": { description: Command acknowledged }
        "404": { description: Command not found }

  # ── Bonus ────────────────────────────────────────────────────────────────────
...
  /api/v1/predictions/{zoneCode}:
    get:
      summary: Advisory ML predicted risk probability (Bonus 3)
      tags: [Bonus]
      parameters:
        - in: path
          name: zoneCode
          required: true
          schema: { type: string }
      responses:
        "200":
          description: ML prediction (advisory only, never actuates hardware)
          content:
            application/json:
              schema:
                type: object
                properties:
                  prediction:
                    type: object
                    properties:
                      probability: { type: number, minimum: 0, maximum: 1 }
                      modelVersion: { type: string }
                      advisoryOnly: { type: boolean, example: true }
                  liveRiskScore: { type: number }
                  safety: { type: string }
  /api/v1/reports/note:
    post:
      summary: Submit natural-language incident report (Bonus 4)
      tags: [Bonus]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  minLength: 5
                  maxLength: 1000
      responses:
        "200":
          description: Report validated and stored
          content:
            application/json:
              schema:
                type: object
                properties:
                  signal:
                    type: object
                    properties:
                      zoneCode: { type: string }
                      hazard: { $ref: "#/components/schemas/HazardType" }
                      severity: { type: string, enum: [LOW, MEDIUM, HIGH] }
                      summary: { type: string }
                  validated: { type: boolean }
                  provider: { type: string, enum: [gemini, openrouter, deterministic] }
                  report_id: { type: string }
                  advisory: { type: string }
        "422": { description: Validation rejected }
        "429": { description: Rate limited }

  # ── Admin ────────────────────────────────────────────────────────────────────
...
  /api/v1/admin/raw-readings:
    get:
      summary: Raw sensor readings — Admin only
      tags: [Admin]
      parameters:
        - in: query
          name: zone
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: limit
          schema: { type: integer, default: 100, maximum: 500 }
      responses:
        "200": { description: Raw readings list }

  # ── System ───────────────────────────────────────────────────────────────────
...